split also test and improve doc
