+++ library/cwm/test/Makefile.am
@@ -1,11 +1,12 @@
 TESTS = \
 	abstract_widget_test.rb \
+	common_widgets_test.rb \
 	custom_widget_test.rb \
 	pager_test.rb \
+	replace_point_test.rb \
 	table_test.rb \
 	tree_test.rb \
-	tree_pager_test.rb \
-	widgets_test.rb
+	tree_pager_test.rb
 
 TEST_EXTENSIONS = .rb
 RB_LOG_COMPILER = rspec
