+++ library/cwm/test/tree_test.rb
